+++ htdocs/_templates/core/_toastv3.php
@@ -1,2 +1,2 @@
 <!-- Toast v3 Notification Container -->
-<div class="toast-panel" id="toast-container"></div>
+<div class="toast-container toast-pos-bottom-right" id="toast-container"></div>